MURF_AI_Voice_Agent/day12_task/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.0
httpx==0.27.0
assemblyai==0.33.0
google-generativeai==0.7.2
jinja2==3.1.4
pydantic==2.6.4
watchfiles==0.22.0
